requirements/base.txt
dj-database-url
Django
python-decouple
pytz
pandas
csv
openpyxl